refactor(core): isolate markdown adapter behind module boundary
Move markdown-specific internals into adapters/markdown/ as private
modules and drop their re-exports from the crate root. SpecInfo,
SpecFrontmatter, SpecStatus, SpecPriority, SpecFilterOptions, and
StatusTransition now live in adapters::markdown::types and are
re-exported only from adapters::markdown. SpecLoader, SpecWriter, and
SpecArchiver become private to the markdown adapter; downstream code
reaches their behaviour through the Adapter trait.

Content string utilities stay generic — re-exported from the crate
root for HTTP handlers using the fetch-transform-push pattern.

Adds MarkdownAdapter::invalidate_path so file-system watchers can hold
a typed adapter handle instead of calling a static cache method.

Per spec #258, leanspec-cli and leanspec-http are intentionally left
broken to serve as a compiler-guided migration map for the downstream
specs (#260+).

Part of #258
