+++ rust/leanspec-core/src/types/mod.rs
@@ -1,9 +1,7 @@
-//! Core types for LeanSpec specs
+//! Core types for LeanSpec
 
 mod config;
-mod spec;
 mod validation;
 
 pub use config::*;
-pub use spec::*;
 pub use validation::*;
